requirements.txt
django==1.7
markdown
djangorestframework
django-filter
pygments

django-celery

